python-ocr/requirements.txt
flask>=3.0.0
pillow>=10.0.0
numpy>=1.24.0
paddleocr>=2.7.0
paddlepaddle>=2.6.0
